console/pyserver/requirements.txt
Django==4.1.3
django-classy-tags==3.0.1
djangorestframework==3.14.0
requests==2.27.1
psycopg2==2.9.5
django-cors-headers==3.13.0
whitenoise==6.2.0
gunicorn==20.1.0
django-inline-svg==0.1.1
phonenumbers==8.13.2
django-phonenumber-field==7.0.1
short_url==1.2.2